Extract UI into its own mixin
