+++ src/mixins/view.js
@@ -6,8 +6,6 @@ import _                        from 'underscore';
 import _getValue                from '../utils/_getValue';
 import getOption                from '../utils/getOption';
 import normalizeMethods         from '../utils/normalizeMethods';
-import normalizeUIKeys          from '../utils/normalizeUIKeys';
-import normalizeUIValues        from '../utils/normalizeUIValues';
 import mergeOptions             from '../utils/mergeOptions';
 import proxyGetOption           from '../utils/proxyGetOption';
 import MarionetteError          from '../error';
@@ -81,21 +79,6 @@ export default {
     return _.extend(target, templateContext);
   },
 
-  // normalize the keys of passed hash with the views `ui` selectors.
-  // `{"@ui.foo": "bar"}`
-  normalizeUIKeys: function(hash) {
-    var uiBindings = _.result(this, '_uiBindings');
-    return normalizeUIKeys(hash, uiBindings || _.result(this, 'ui'));
-  },
-
-  // normalize the values of passed hash with the views `ui` selectors.
-  // `{foo: "@ui.bar"}`
-  normalizeUIValues: function(hash, properties) {
-    var ui = _.result(this, 'ui');
-    var uiBindings = _.result(this, '_uiBindings');
-    return normalizeUIValues(hash, uiBindings || ui, properties);
-  },
-
   // Configure `triggers` to forward DOM events to view
   // events. `triggers: {"click .foo": "do:foo"}`
   configureTriggers: function() {
@@ -129,8 +112,8 @@ export default {
     var combinedEvents = _.extend({},
       this._getBehaviorEvents(),
       events,
-      this.configureTriggers(),
-      this._getBehaviorTriggers()
+      this._getBehaviorTriggers(),
+      this.configureTriggers()
     );
 
     Backbone.View.prototype.delegateEvents.call(this, combinedEvents);
@@ -206,55 +189,10 @@ export default {
     this._bindBehaviorUIElements();
   },
 
-  // This method binds the elements specified in the "ui" hash inside the view's code with
-  // the associated jQuery selectors.
-  _bindUIElements: function() {
-    if (!this.ui) { return; }
-
-    // store the ui hash in _uiBindings so they can be reset later
-    // and so re-rendering the view will be able to find the bindings
-    if (!this._uiBindings) {
-      this._uiBindings = this.ui;
-    }
-
-    // get the bindings result, as a function or otherwise
-    var bindings = _.result(this, '_uiBindings');
-
-    // empty the ui so we don't have anything to start with
-    this._ui = {};
-
-    // bind each of the selectors
-    _.each(bindings, function(selector, key) {
-      this._ui[key] = this.$(selector);
-    }, this);
-
-    this.ui = this._ui;
-  },
-
   // This method unbinds the elements specified in the "ui" hash
   unbindUIElements: function() {
     this._unbindUIElements();
     this._unbindBehaviorUIElements();
-  },
-
-  _unbindUIElements: function() {
-    if (!this.ui || !this._uiBindings) { return; }
-
-    // delete all of the existing ui bindings
-    _.each(this.ui, function($el, name) {
-      delete this.ui[name];
-    }, this);
-
-    // reset the ui element to the original bindings configuration
-    this.ui = this._uiBindings;
-    delete this._uiBindings;
-    delete this._ui;
-  },
-
-  getUI: function(name) {
-    this._ensureViewIsIntact();
-
-    return this._ui[name];
   },
 
   // Internal method to create an event handler for a given `triggerDef` like
